backend/requirements.txt
fastapi
uvicorn
authlib
psycopg2-binary
python-dotenv
passlib[bcrypt]
redis